requirements.txt
attrs==23.2.0
aws-cdk-lib==2.149.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
boto3==1.34.158
botocore==1.34.158
cattrs==23.2.3
constructs==10.3.0
importlib_resources==6.4.0
jmespath==1.0.1
jsii==1.101.0
publication==0.0.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.2
